Rename custom widget to customapi
